Moved Main menu class to own file, tidied up
Moved Main menu class to own file, wrote tests for pinging api, tidied up cryptoapi.rb
