+++ .gitignore
@@ -9,6 +9,6 @@ tmp
 *.iml
 *.log
 *.swp
+*.json
 
 db/*.db
-*.json
