+++ .results.json
@@ -1,1 +1,1 @@
-{"version":"3.11.0","examples":[{"id":"./spec/models/Crypto_Api_spec.rb[1:1:1]","description":"Should show Success for a successful ping","full_description":"CryptoAPI#ping_api Should show Success for a successful ping","status":"passed","file_path":"./spec/models/Crypto_Api_spec.rb","line_number":8,"run_time":0.005736,"pending_message":null}],"summary":{"duration":0.006774,"example_count":1,"failure_count":0,"pending_count":0,"errors_outside_of_examples_count":0},"summary_line":"1 example, 0 failures"}
+{"version":"3.11.0","examples":[{"id":"./spec/models/Crypto_Api_spec.rb[1:1:1]","description":"On successful ping should set connection status hash object message to \"Connected to API\"","full_description":"CryptoAPI#ping_api On successful ping should set connection status hash object message to \"Connected to API\"","status":"passed","file_path":"./spec/models/Crypto_Api_spec.rb","line_number":9,"run_time":0.005691,"pending_message":null},{"id":"./spec/models/Crypto_Api_spec.rb[1:1:2]","description":"On bad_request ping should set connection status hash object message to \"Unable to connect to API, Please close and reopen.\"","full_description":"CryptoAPI#ping_api On bad_request ping should set connection status hash object message to \"Unable to connect to API, Please close and reopen.\"","status":"passed","file_path":"./spec/models/Crypto_Api_spec.rb","line_number":14,"run_time":0.000187,"pending_message":null}],"summary":{"duration":0.006644,"example_count":2,"failure_count":0,"pending_count":0,"errors_outside_of_examples_count":0},"summary_line":"2 examples, 0 failures"}
