+++ bin/run.rb
@@ -1,10 +1,8 @@
 require_relative '../config/environment'
-require_relative '../lib/run'
+require_relative '../lib/models/Main_Menu.rb'
 
 menu = Main_Menu.new
 menu.startup
 while !$quit
     menu.display_main_menu
 end
-# loading_menu
-#display_main_menu
